x % Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractCluster.R
\name{AbstractCluster}
\alias{AbstractCluster}
\title{Implements}
\usage{
AbstractCluster(dat, distance = 10)
}
\arguments{
\item{dat}{A data set}

\item{distance}{The radius around a center point where the function will look for points to add to it's cluster}
}
\value{
Returns cluster assignments
}
\description{
Implements
}
