% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractCluster.R
\name{ClusterPlot}
\alias{ClusterPlot}
\title{Cluster plot}
\usage{
ClusterPlot(abstract_clusters)
}
\arguments{
\item{clusters}{Clusters from AbstractCluster}
}
\value{
plot
}
\description{
Cluster plot
}
