% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractCluster.R
\name{CreateElbowGraph}
\alias{CreateElbowGraph}
\title{Creates an Elbow Method Graph to help the user manually choose # of clusters necessary for their k-means calculation.}
\usage{
CreateElbowGraph(dat, explanatory, response, kmax)
}
\arguments{
\item{dat}{A data set}

\item{explanatory}{explanatory variable of interest}

\item{response}{response variable of interest}

\item{kmax}{the maximum number of clusters the program should run and graph}
}
\value{
Elbow Method Graph
}
\description{
Creates an Elbow Method Graph to help the user manually choose # of clusters necessary for their k-means calculation.
}
