% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractCluster.R
\name{getWCSS}
\alias{getWCSS}
\title{Helper function that calculates the within cluster sum of squares}
\usage{
getWCSS(clusters)
}
\arguments{
\item{clusters}{a dataframe that represents a cluster}
}
\description{
Helper function that calculates the within cluster sum of squares
}
