% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractCluster.R
\name{AbstractCluster}
\alias{AbstractCluster}
\title{Implements}
\usage{
AbstractCluster(dat, distance = 0.1, neighbors = 5)
}
\arguments{
\item{dat}{A data set}

\item{k}{The number of desired clusters}
}
\value{
Returns cluster assignments
}
\description{
Implements
}
