% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractCluster.R
\name{AbstractCluster}
\alias{AbstractCluster}
\title{Forms clusters by picking a random observation and finding points that are within a specified distance to that center point}
\usage{
AbstractCluster(dat, distance = 4)
}
\arguments{
\item{dat}{A data set}

\item{distance}{The radius around a center point where the function will look for points to add to it's cluster}
}
\value{
Returns cluster assignments
}
\description{
Forms clusters by picking a random observation and finding points that are within a specified distance to that center point
}
