% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractCluster.R
\name{getWCSS}
\alias{getWCSS}
\title{Helper function for CreateElbowGraph that calculates the within cluster sum of squares}
\usage{
getWCSS(clusters)
}
\arguments{
\item{clusters}{a data frame that represents a cluster}
}
\value{
integer describing a cluster's sum of squares from the centroid
}
\description{
Helper function for CreateElbowGraph that calculates the within cluster sum of squares
}
